requirements.txt
flask
requests
python-owasp-zap-v2.4